Unbreak on 32-bit architectures

+++ src/backend.cpp
@@ -167,7 +167,7 @@ void X11Client::UpdateTranslation(){
 		
 		if(pcontainer->pParent->flags & WManager::Container::FLAG_STACKED){
 			uint stackIndex = m-pcontainer->pParent->stackQueue.begin();
-			float stackOffset = 1.0f/(float)std::max(pcontainer->pParent->stackQueue.size(),1lu);
+			float stackOffset = 1.0f/(float)std::max((unsigned long)pcontainer->pParent->stackQueue.size(),1lu);
 			pcontainer->titleSpan = glm::vec2((float)stackIndex*stackOffset,((float)stackIndex+1.0f)*stackOffset);
 
 			titleFrameExtent = glm::vec2(coord.z,coord.w)*stackOffset;
@@ -1504,7 +1504,7 @@ void DebugClient::UpdateTranslation(){
 		if(pcontainer->pParent->flags & WManager::Container::FLAG_STACKED){
 			uint stackIndex = m-pcontainer->pParent->stackQueue.begin();
 			titleFrameExtent = glm::vec2(coord.z,coord.w)
-				/(float)std::max(pcontainer->pParent->stackQueue.size(),1lu);
+				/(float)std::max((unsigned long)pcontainer->pParent->stackQueue.size(),1lu);
 			titleStackOffset = (float)stackIndex*titleFrameExtent;
 		}else{
 			titleFrameExtent = glm::vec2(coord.z,coord.w);
